+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.1.0]
+
+### [Add edit functionality]
+
+#### Added
+
+* Add Edit behaviour
+* Add Select all button
+* Add deletation functionality
+* Add sharing functionality
+
 ## [1.0.0]
 
 ### [Add web server integration]